Use diffrent dummy data for PAAPI and prebid
